Extracted `RectExtensions` to discrete script
