+++ Editor/Extensions/RectExtensions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a7c2026af8e7467429f67a4ed2b7caef
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
